tegra: common: move to DT based DAI links
Current 'sound' node does not describe all the DAI links that Tegra sound
card uses. Most of the links are statically declared in the driver code
and only few codec links are exposed from the DT. As the plan is to move
towards DT based binding for all DAI links, the sound node is removed
from 'tegra-platforms-audio-dai-links.dtsi'. This file currently enables
audio nodes of APE/AHUB and populates plugin related info for ADSP. Going
ahead this file either needs to be removed or renamed for better clarity
with naming of files. Though it is present in common path, it is used on
Tegra194 platforms only.

As part of the new approach following dtsi files are proposed.
* tegra210-audio-dai-links.dtsi : This includes all the DAI links needed
  for Tegra210 based platforms.
* tegra186-audio-dai-links.dtsi : This includes all the DAI links needed
  for Tegra186 based platforms. Since AHUB architecture of Tegra194 is
  similar to Tegra186, platforms based on Tegra194 can use the same dtsi
  file.

This patch also updates audio routing property name to be similar in style
with upstream DT binding for simple card.

Bug 200538260

Change-Id: If80b1de9baea572882c6ffb37d9a643d224c7088
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2209424
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>