dt-bindings: tegra: add DAI index header
This patch adds macros for all the DAIs used on Tegra210 and Tegra186
based platforms. The macro values are closely tied with the DAI arrays
defined in AHUB drivers. Indices are combined for both Tegra210 and
Tegra186. No need to maintain separate macros for both and duplication
is avoided by following ways for XBAR.
 * List out common DAIs in the beginning.
 * List out Tegra210 specific DAIs after common list and then followed
   by Tegra186 DAIs.

For ADMAIF indices are listed for all channels. Only a subset of these
are used on Tegra210. For other modules, including ADSP, DAI indices
are common and does not require specific handling.

Bug 200538260

Change-Id: Id1956d4c10efff35f37d1bb3bf5abd12772d8188
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2209427
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>