video: tegra: expose AVI rgb_ycc control to user-space
In certain conditions, sink requires that the source signals a
specific encoding of pixels transmitted over physical link,
regardless the actual encoding of transmitted pixels.  For
instance, a Dolby Vision compliant sinks may require that
pixels are encoded as YUV422, but the link is setup as RGB - so
called tunneled mode.

This change enables to setup the signaling of specific encoding
of pixels from user-space.

Bug 200555042

Change-Id: I2355cecb43e11d22eec7cc08269e732410e08e36
Signed-off-by: Michael Frydrych <mfrydrych@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2204903
Reviewed-on: https://git-master.nvidia.com/r/2255102
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>