safety: mc: return 0xffff to suppress errors
MC Code returned 0x0 when the driver
reads MC register space in safety build.
Number of memory channels is actually
determined at runtime by reading a register
and returning a 0 here lead to a lot of warnings
and errors during boot time.
Hence this CL makes the change to return 0xfff
which signifies that the number of memory
channels is 16.

JIRA SSV-3514

Change-Id: I41fec870d8d7761dc456815608dc7194a479812e
Signed-off-by: at <at@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2218692
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Nitin Sehgal <nsehgal@nvidia.com>
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>