t19x: galen: Fix audio route for FE-PI Audio Z V2
The audio routing for the FE-PI Audio Z V2 board is different on Jetson
Xavier that other Jetson platforms because it is not the primary I2S
interface that is used for the codec and so the codec prefix is 'y' and
not 'x'. Fix this by overriding the audio route for Jetson Xavier in the
top-level device-tree source file for the FE-PI Audio Z V2 on Jetson
Xavier.

Bug 2414131

Change-Id: Id4e57995ae7dfc5546514b653f90099e3ce084e0
Reviewed-on: https://git-master.nvidia.com/r/2220851
(cherry picked from commit 351823a0aeb617c137d1764ad14d7eeda9b84f1b)
Reviewed-on: https://git-master.nvidia.com/r/2224178
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Tested-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>