t19x: jakku: Add overlays for Jetson-Xavier-NX
Add the following device-tree overlays for the Jetson Xavier-NX
platform.

1. ADAFRUIT SPH0645LM4H (I2S MEMS Microphone)
2. FE-PI Audio Z V2 (SGTL5000 Audio Codec)
3. Jetson 40-pin Expansion Header

The Jetson 40-pin Expansion Header overlay has been developed for
the 'Jetson-IO' tool to help user re-configure the IOs exposed by
the header. This overlay adds a 'header-40pin-pinmux' node under
the pinmux device-node that defines the mapping of the SoC IOs to
the pins on the header. It also defines a default configuration for
the pins exposed by the header.

The overlays for the ADAFRUIT and FE-PI add-on modules are developed
to be directly applied to the base DTB file for the Jetson platform.
For example, these overlays can be used a follows in order to generate
a complete DTB with the overlay applied.

 $ fdtoverlay -i <base-dtb> -o <out-dtb> <overlay-dtbo>

Finally, add phandles for nodes that are common to the 40-pin header,
for example the 3.3V and 5V supplies, so that overlays can use these
phandles and work all the Jetson platforms.

Bug 200564376

Change-Id: Ib6ae73f76bdd42b8284b485f9022bfa0c023c4a2
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2237660
(cherry picked from commit c2284d88953f29bfbb93f656f8e88b343d5f8dd0)
Reviewed-on: https://git-master.nvidia.com/r/2240044
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>