ASoC: tegra-alt: simplify timeout code for I2S and ADMAIF
For SW reset and register status polling timeout code is used in multiple
drivers. Since the logic is similar, the same code is duplicated at many
places. This can be avoided by relying on the kernel macro which provides
similar functionality, where sleep between polling and timeout can be
explicitly provided. For example, ADMA driver is using a similar kernel
macro and this can be used for Tegra ASoC drivers as well.

Currently this patch updates timeout code for ADMAIF and I2S to start with.
This helps for upstream as well. Other drivers will be taken care in a
separate patch. In doing so following changes are also needed,
 * Few functions are optimized to simplify the code during stream start or
   stop. This basically combines RX and TX handling in a single function
   and thus helps to avoid code duplication.
 * For ADMAIF, macros are used wherever register access for RX and TX paths
   are needed. Existing CH_REG() macro is updated to facilitate this. RX
   and TX paths can use CH_RX_REG() and CH_TX_REG() respectively to get
   proper register offset. Accordingly the access is updated in the driver.

Bug 200566596
Bug 200503387

Change-Id: I6d46518dba09a936ad84799943eba8976a261faa
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2241448
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>