drivers:media: camera_common
The resolutions for sensors which have
use_sensor_mode_id defined in the DT should
be checked against the available mode sizes.
This was exposed when we removed width alignment
with bytesperline in tegra channel. This was
causing the VIDIOC_TRY_FMT and VIDIOC_S_FMT
ioctls to fail in v4l2 compliance.

bug 200569152

Change-Id: If46b5c99e8a6ce08f29dc03c2a9fc3efacc0499c
Signed-off-by: Anubhav Rai <arai@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2249692
Reviewed-by: Frank Chen <frankc@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Winnie Hsu <whsu@nvidia.com>
Reviewed-by: Sean Pieper <spieper@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2252680
Reviewed-by: Sudhir Vyas <svyas@nvidia.com>