t194: Update upstream iommu DT bindings
Since kernel version 5.5, "dma-ranges" property needs to
be very carefully used as the kernel probes this property
in a different direction than before. And this causes our
downstream stable tree having OF errors like:
 OF: translation of DMA address(0) to CPU address failed node(...)

This is because our downstream DT binding has dma-ranges
in its root node so fails the probing logic. A simple fix
is to remove the dma-ranges properties, in root node and
host1x.

As these two properties were used to limit IOVA addresses
space, there needs to be an alternative solution after we
remove them.

So this patch also introduces the upstream solution. It
changes memory-controller node as an interconnects type
which has the dma-ranges property to limit the IOVA, and
adds to clients corresponding interconnects properties.
This allows all the client devices to have bus_dma_limit
in their dev pointers so that this bus_dma_limit will be
applied to the dma_limit whenever allocating a new IOVA.

Bug 200577906

Change-Id: Id14cd515e4fd0af6251491dac668b32918ec92cd
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t19x/+/2279177
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>